backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
pynacl>=1.5.0
cryptography>=42.0
argon2-cffi>=23.1
PyJWT[crypto]>=2.8
pydantic-settings>=2.0
keyring>=25.0
websockets>=12.0
httpx>=0.27
slowapi>=0.1.9
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
python-multipart>=0.0.9
